Major shift to use a Future style interface to enable better async style behaviour and FRP composition during service orchestration/ pipelining.

+++ muon-tck/src/main/java/com/simplicity/services/ServicePublishColdStream.java
@@ -24,14 +24,14 @@ public class ServicePublishColdStream {
         new AmqpTransportExtension("amqp://localhost").extend(muon);
         muon.start();
 
-        muon.streamSource("/counter", Integer.class, new MuonStreamGenerator<Integer>() {
+        muon.streamSource("/counter", Long.class, new MuonStreamGenerator<Long>() {
             @Override
-            public Publisher<Integer> generatePublisher(Map<String, String> parameters) {
-                int max = Integer.parseInt(parameters.get("max"));
+            public Publisher<Long> generatePublisher(Map<String, String> parameters) {
+                long max = Long.parseLong(parameters.get("max"));
                 return Streams.range(0, max);
             }
         });
 
-        muon.streamSource("/countersimple", Integer.class, Streams.range(0, 100));
+        muon.streamSource("/countersimple", Long.class, Streams.range(0, 100));
     }
 }
